Hardware Patch
Since Pin D13 is raised HIGH during bootup due to the arduino bootloader, the relay turns on for a brief moment during boot-up. This is unsafe for the purpose of this device.

So the relay control pin was moved from D13 to D12. Picture of the hardware patch has been included in the github readme